refactor: transform utility classes to objects
