+++ owncloudApp/src/main/java/com/owncloud/android/presentation/documentsprovider/DocumentsProviderUtils.kt
@@ -23,15 +23,13 @@ import android.content.Context
 import android.provider.DocumentsContract
 import com.owncloud.android.R
 
-class DocumentsProviderUtils {
-    companion object {
-        /**
-         * Notify Document Provider to refresh roots
-         */
-        fun notifyDocumentsProviderRoots(context: Context) {
-            val authority = context.resources.getString(R.string.document_provider_authority)
-            val rootsUri = DocumentsContract.buildRootsUri(authority)
-            context.contentResolver.notifyChange(rootsUri, null)
-        }
+object DocumentsProviderUtils {
+    /**
+     * Notify Document Provider to refresh roots
+     */
+    fun notifyDocumentsProviderRoots(context: Context) {
+        val authority = context.resources.getString(R.string.document_provider_authority)
+        val rootsUri = DocumentsContract.buildRootsUri(authority)
+        context.contentResolver.notifyChange(rootsUri, null)
     }
 }
